requirements.txt
catboost==0.23
click==7.1.2
cycler==0.10.0
future==0.18.2
graphviz==0.14
grpcio==1.28.1
grpcio-tools==1.28.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.5.2
matplotlib==3.2.1
numpy==1.18.3
pandas==1.0.3
plotly==4.6.0
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torch==1.5.0
Werkzeug==1.0.1
